Date and amount
To reduce the amount of columns per row, I have removed the UUID portion
of the table. However, I have added a start date column and an end date
column that convert the timestamp from the database into a real date and
time. Along with this, the buttons for each punishment type now show the
amount of punishments for that type.