User profile and new search
When searching based on a specific user, a visitor will now be sent to a
seperate page that contains a full 3D skin model of the player and just
their punishments. Because it is their personal page, no punishment type
variations are provided in the header and no usernames are required for
the table of punishments.