Banned badge
As requested by a few users, I have added a badge under the user's name
to display if the user is banned, temporarily banned, or not banned at
all. With this, I have also moved the profile portion of the user page
to the right hand side. However, with this update, there have been some
loading problems. It seems to take the page quite a lot longer to load
than usual. I hope to fix this soon.